chalice now can load views from files
